backend/docker/requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
Pillow==10.4.0
reportlab==4.2.2
boto3==1.34.144
